requirements.txt
Flask==3.0.0
tensorflow==2.17.1
numpy==1.26.0
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
matplotlib==3.8.0